requirements.txt
backoff
datasets
docopt
hydra-core
lexrank
pandas==1.3.0
pytorch-lightning==1.3.8
p_tqdm
tqdm
scipy==1.7.0
scispacy==0.4.0
nltk
jsonlines
ratelimit
torch==1.8.1+cu101 -f https://download.pytorch.org/whl/cu101/torch_stable.html
transformers==4.8.2
ujson
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.4.0/en_core_sci_sm-0.4.0.tar.gz
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.3.0/en_core_sci_lg-0.3.0.tar.gz
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.4.0/en_core_sci_md-0.4.0.tar.gz
